+++ BazarCarioca.WebAPI/Controllers/StoresController.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using BazarCarioca.WebAPI.DTOs;
+using BazarCarioca.WebAPI.Extensions;
 using BazarCarioca.WebAPI.Models;
 using BazarCarioca.WebAPI.Repositories;
 using BazarCarioca.WebAPI.Services;
@@ -25,32 +26,44 @@ namespace BazarCarioca.WebAPI.Controllers
         }
 
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Store>>> Get()
+        public async Task<IActionResult> Get()
         {
             var stores = await Repository.GetAsync();
+
+            if (stores.IsNullOrEmpty())
+                return NotFound("Nenhuma loja foi encontrada.");
 
             return Ok(stores);
         }
 
         [HttpGet("{Id:int}")]
-        public async Task<ActionResult<Store>> GetById(int Id)
+        public async Task<IActionResult> GetById(int Id)
         {
             var store = await Repository.GetByIdAsync(Id);
+            
+            if (store == null)
+                return NotFound($"A loja com Id = {Id} não foi encontrada.");
 
             return Ok(store);
         }
 
         [HttpGet("Lojista/{Id:int}")]
-        public async Task<ActionResult<IEnumerable<Store>>> GetStoresByShopkeeperId(int Id)
+        public async Task<IActionResult> GetStoresByShopkeeperId(int Id)
         {
             var stores = await Repository.GetByShopkeeperIdAsync(Id);
+
+            if (stores == null)
+                return NotFound("Lojista inexiste ou sem lojas.");
 
             return Ok(stores);
         }
 
         [HttpPost]
-        public async Task<ActionResult<StoreDTO>> Create([FromForm] StoreCreateDTO createDto)
+        public async Task<IActionResult> Create([FromForm] StoreCreateDTO createDto)
         {
+            if (createDto == null)
+                return BadRequest("Nenhuma informação foi passada na requisição.");
+
             var store = Mapper.Map<Store>(createDto);
 
             if (createDto.File != null)
@@ -64,17 +77,14 @@ namespace BazarCarioca.WebAPI.Controllers
         }
 
         //Só funciona com Postman
-        //Refinar lógica principalmente com WebService
         [HttpPatch("{Id:int}")]
         [Consumes("multipart/form-data")]
-        public async Task<ActionResult<StoreDTO>> Patch(int Id, [FromForm] PatchRequestDTO requestDto)
+        public async Task<IActionResult> Patch(int Id, [FromForm] PatchRequestDTO requestDto)
         {
-            var productPatched = await Repository.UpdateWithImageAsync(Id, requestDto);
+            if (requestDto == null)
+                return BadRequest("Houve um erro na requisição HTTP. Informações não foram enviadas.");
 
-            if (requestDto.File != null)
-                Console.WriteLine("***** Pelo visto NÂO é null");
-            else
-                Console.WriteLine("***** Pelo visto é null");
+            var productPatched = await Repository.UpdateWithImageAsync(Id, requestDto);
 
             var productDto = Mapper.Map<StoreDTO>(productPatched);
 
@@ -82,11 +92,21 @@ namespace BazarCarioca.WebAPI.Controllers
         }
 
         [HttpDelete("{Id:int}")]
-        public async Task<ActionResult<bool>> DeleteStore(int Id)
+        public async Task<IActionResult> DeleteStore(int Id)
         {
+            var store = await Repository.GetByIdAsync(Id);
+
+            if (store == null)
+                return NotFound($"A loja não foi apagada pois não existe uma loja com Id = {Id}.");
+
+            var fileUrl = store.ImageUrl;
+
+            if (fileUrl != null)
+                await WebService.DeleteFileAsync(fileUrl);
+
             await Repository.DeleteAsync(Id);
 
-            return Ok($"Loja com id = {Id} apagada.");
+            return Ok($"A loja com id = {Id} foi apagada.");
         }
     }
 }
